+++ src/index.ts
@@ -10,3 +10,4 @@ export * from "./functions/generate-random-string.js";
 export * from "./functions/is-local-ip-address.js";
 export * from "./functions/pad-string-with-null.js";
 export * from "./functions/split-string-into-chunks.js";
+export * from "./functions/string-replace-all-async.js";